Finished basic GLSL functionality, see description
I think I did a pretty good job distinguishing between what should be in
the container and what should be in the static utility classes.  This is
not a final revision, and you are welcome to tweak javadocs, methods,
and whatever!  If you don't like how it wraps GLSL functionality, let me
know what's wrong with it.