Added lwjgl as test dependency (see below) and started deployment
jgloom-io has jgloom-lwjgl as a TEST dependency, so only tests can access it (which is required for actual GL testing).  Main Java code cannot access jgloom-lwjgl, which is good.  Deployment code is having the kinks worked out and should be done soon.  DO NOT REMOVE GRADLE.PROPERTIES FROM THE .gitignore, IT CONTAINS PASSWORDS IN PLAIN TEXT!